@@ -1,19 +1,13 @@
 {
   "name": "ideaa/cross",
   "type": "library",
-  "description": "Simple php5 Framework",
+  "description": "light php framework",
   "keywords": [
     "cross",
     "crossphp",
-    "php5 Framework"
+    "PHP Framework"
   ],
   "homepage": "http://www.crossphp.com",
-  "repositories": [
-    {
-      "type": "vcs",
-      "url": "https://github.com/wonli/crossphp.git"
-    }
-  ],
   "license": "MIT",
   "authors": [
     {
@@ -23,22 +17,23 @@
       "role": "Developer"
     }
   ],
+  "autoload": {
+    "psr-4": {
+      "Cross\\": "/src"
+    }
+  },
   "require": {
-    "php": ">=5.3.6",
+    "php": "^7.2.0",
     "ext-json": "*",
     "ext-dom": "*",
     "ext-pdo": "*",
     "ext-mbstring": "*",
     "ext-iconv": "*",
     "ext-curl": "*",
-    "ext-openssl": "*"
+    "ext-openssl": "*",
+    "ext-zlib": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "4.3.*"
-  },
-  "autoload": {
-    "psr-4": {
-      "Cross\\": "/src"
-    }
+    "phpunit/phpunit": "^9"
   }
 }
