@@ -1,6 +1,6 @@
 <?php
 /**
- * Cross - a micro PHP 5 framework
+ * Cross - a micro PHP framework
  *
  * @link        http://www.crossphp.com
  * @license     MIT License
@@ -51,7 +51,7 @@ class Annotate
      * @param Delegate $delegate
      * @return Annotate
      */
-    public static function getInstance(Delegate &$delegate)
+    public static function getInstance(Delegate &$delegate): self
     {
         if (!self::$instance) {
             self::$instance = new Annotate($delegate);
@@ -66,7 +66,7 @@ class Annotate
      * @param string $prefix
      * @return $this
      */
-    function setPrefix($prefix)
+    function setPrefix(string $prefix): self
     {
         $this->prefix = $prefix;
         return $this;
@@ -78,18 +78,28 @@ class Annotate
      * @param string $annotate
      * @return array
      */
-    public function parse($annotate = '')
+    public function parse(string $annotate = ''): array
     {
         if (empty($annotate)) {
-            return array();
+            return [];
         }
 
         $flag = preg_match_all("/@{$this->prefix}(.*?)\s+(.*)/", $annotate, $content);
-        if (!$flag) {
-            return array();
+        if (!$flag || empty($content[1])) {
+            return [];
         }
 
-        $configs = array_combine($content[1], $content[2]);
+        $configs = [];
+        $values = &$content[2];
+        array_walk($content[1], function ($k, $index) use ($values, &$configs) {
+            $v = &$values[$index];
+            if (isset($configs[$k])) {
+                $configs[$k] .= "\n" . $v;
+            } else {
+                $configs[$k] = $v;
+            }
+        });
+
         return $this->parseAnnotate($configs);
     }
 
@@ -99,7 +109,7 @@ class Annotate
      * @param string $params
      * @return Closure
      */
-    public function bindToClosure($params)
+    public function bindToClosure(string $params): Closure
     {
         return function ($self) use ($params) {
             return include("annotate://{$params}");
@@ -112,7 +122,7 @@ class Annotate
      * @param string $params
      * @return mixed
      */
-    public function toCode($params)
+    public function toCode(string $params)
     {
         return include("annotate://{$params}");
     }
@@ -128,12 +138,12 @@ class Annotate
      * )
      * </pre>
      *
-     * @param $params
+     * @param string $params
      * @return array
      */
-    public function toArray($params)
+    public function toArray(string $params): array
     {
-        $result = array();
+        $result = [];
         $conf = array_map('trim', explode(',', $params));
         foreach ($conf as $c) {
             if (false !== strpos($c, '=')) {
@@ -144,7 +154,6 @@ class Annotate
             }
         }
         unset($conf);
-
         return $result;
     }
 
@@ -154,9 +163,9 @@ class Annotate
      * @param array $annotateConfigs
      * @return array
      */
-    private function parseAnnotate(array $annotateConfigs)
+    private function parseAnnotate(array $annotateConfigs): array
     {
-        $result = array();
+        $result = [];
         foreach ($annotateConfigs as $conf => $params) {
             switch ($conf) {
                 case 'params':
