@@ -1,6 +1,6 @@
 <?php
 /**
- * Cross - a micro PHP 5 framework
+ * Cross - a micro PHP framework
  *
  * @link        http://www.crossphp.com
  * @license     MIT License
@@ -8,8 +8,11 @@
 
 namespace Cross\Core;
 
-use Cross\Http\Request;
+use Cross\Exception\CoreException;
+
 use DOMDocument;
+use Exception;
+use DateTime;
 
 /**
  * @author wonli <wonli@live.com>
@@ -22,11 +25,11 @@ class Helper
      * 截取字符串
      *
      * @param string $str 要截取的字符串参数
-     * @param string $len 截取的长度
+     * @param int $len 截取的长度
      * @param string $enc 字符串编码
      * @return string
      */
-    public static function subStr($str, $len, $enc = 'utf8')
+    public static function subStr(string $str, int $len, string $enc = 'utf8'): string
     {
         if (self::strLen($str) > $len) {
             return mb_substr($str, 0, $len, $enc) . '...';
@@ -39,11 +42,11 @@ class Helper
      * 安全的截取HTML字符串
      *
      * @param string $str 要截取的字符串参数
-     * @param string $len 截取的长度
+     * @param int $len 截取的长度
      * @param string $enc 字符串编码
      * @return string
      */
-    public static function subStrHTML($str, $len, $enc = 'utf8')
+    public static function subStrHTML(string $str, int $len, string $enc = 'utf8'): string
     {
         $str = self::subStr($str, $len, $enc);
         return self::formatHTMLString($str);
@@ -56,7 +59,7 @@ class Helper
      * @param bool $removing_doctype
      * @return string
      */
-    public static function formatHTMLString($str, $removing_doctype = true)
+    public static function formatHTMLString(string $str, bool $removing_doctype = true): string
     {
         $DOCUMENT = new DOMDocument();
         @$DOCUMENT->loadHTML(mb_convert_encoding($str, 'HTML-ENTITIES', 'UTF-8'));
@@ -75,9 +78,9 @@ class Helper
      * @param string $enc 默认utf8编码
      * @return int
      */
-    public static function strLen($str, $enc = 'gb2312')
+    public static function strLen(string $str, string $enc = 'gb2312'): int
     {
-        return min(array(mb_strlen($str, $enc), mb_strlen($str, 'utf-8')));
+        return min([mb_strlen($str, $enc), mb_strlen($str, 'utf-8')]);
     }
 
     /**
@@ -87,13 +90,13 @@ class Helper
      * @param string $charset 字符编码 默认utf-8
      * @return array
      */
-    static function stringToArray($str, $charset = 'utf-8')
+    static function stringToArray(string $str, string $charset = 'utf-8'): array
     {
         if ($charset != 'utf-8') {
             $str = iconv($charset, 'utf-8', $str);
         }
 
-        $result = array();
+        $result = [];
         for ($i = 0, $str_len = mb_strlen($str, 'utf-8'); $i < $str_len; $i++) {
             $result[] = mb_substr($str, $i, 1, 'utf-8');
         }
@@ -107,7 +110,7 @@ class Helper
      * @param string $str
      * @return string
      */
-    static function md10($str = '')
+    static function md10($str = ''): string
     {
         return substr(md5($str), 10, 10);
     }
@@ -118,7 +121,7 @@ class Helper
      * @param string $file 文件名
      * @return string
      */
-    static function getExt($file)
+    static function getExt(string $file): string
     {
         return pathinfo($file, PATHINFO_EXTENSION);
     }
@@ -131,7 +134,7 @@ class Helper
      * @param bool $recursive
      * @return bool
      */
-    static function createFolders($path, $mode = 0755, $recursive = true)
+    static function createFolders(string $path, int $mode = 0755, bool $recursive = true): bool
     {
         if (!is_dir($path)) {
             return mkdir($path, $mode, $recursive);
@@ -148,7 +151,7 @@ class Helper
      * @param int $dir_mode
      * @return bool
      */
-    static function mkfile($file_name, $mode = 0644, $dir_mode = 0755)
+    static function mkfile(string $file_name, int $mode = 0644, int $dir_mode = 0755): bool
     {
         if (!file_exists($file_name)) {
             $file_path = dirname($file_name);
@@ -174,7 +177,7 @@ class Helper
      * @param string $add_valid_expr
      * @return bool
      */
-    static function validEmail($email, $add_valid_expr = "/^[a-zA-Z0-9]([\w\-\.]?)+/")
+    static function validEmail(string $email, string $add_valid_expr = "/^[a-zA-Z0-9]([\w\-\.]?)+/"): bool
     {
         if (filter_var($email, FILTER_VALIDATE_EMAIL)) {
 
@@ -198,9 +201,9 @@ class Helper
      * @param int $numeric
      * @return string
      */
-    static function random($length, $numeric = 0)
+    static function random(int $length, int $numeric = 0): string
     {
-        $seed = md5(print_r($_SERVER, 1) . microtime(true));
+        $seed = md5(microtime(true));
         if ($numeric) {
             $seed = str_replace('0', '', base_convert($seed, 16, 10)) . '0123456789';
         } else {
@@ -222,7 +225,7 @@ class Helper
      * @param string $str
      * @return array
      */
-    static function parseAt($str)
+    static function parseAt(string $str): array
     {
         preg_match_all("/@([^@^\\s^:]{1,})([\\s\\:\\,\\;]{0,1})/", $str, $result);
         return $result;
@@ -235,10 +238,10 @@ class Helper
      * @param string $disallowable
      * @return mixed
      */
-    static function stripSelectedTags($str, $disallowable = '<script><iframe><style><link>')
+    static function stripSelectedTags(string $str, string $disallowable = '<script><iframe><style><link>'): string
     {
-        $disallowable = trim(str_replace(array('>', '<'), array('', '|'), $disallowable), '|');
-        $str = str_replace(array('&lt;', '&gt;'), array('<', '>'), $str);
+        $disallowable = trim(str_replace(['>', '<'], ['', '|'], $disallowable), '|');
+        $str = str_replace(['&lt;', '&gt;'], ['<', '>'], $str);
         $str = preg_replace("~<({$disallowable})[^>]*>(.*?<\s*\/(\\1)[^>]*>)?~is", '$2', $str);
 
         return $str;
@@ -250,9 +253,9 @@ class Helper
      * @param string $str
      * @return string
      */
-    static function convertTags($str)
+    static function convertTags(string $str): string
     {
-        return str_replace(array('<', '>', "'", '"'), array('&lt;', '&gt;', '&#039;', '&quot;'), $str);
+        return str_replace(['<', '>', "'", '"'], ['&lt;', '&gt;', '&#039;', '&quot;'], $str);
     }
 
     /**
@@ -264,7 +267,7 @@ class Helper
      * @param int $expiry
      * @return string
      */
-    static function authCode($string, $operation = 'DECODE', $key = 'crossphp', $expiry = 0)
+    static function authCode(string $string, string $operation = 'DECODE', string $key = 'crossphp', int $expiry = 0): string
     {
         $c_key_length = 4;
         $key = md5($key);
@@ -281,11 +284,11 @@ class Helper
             base64_decode(substr($string, $c_key_length)) :
             sprintf('%010d', $expiry ? $expiry + time() : 0) . substr(md5($string . $key_b), 0, 16) . $string;
 
-        $result = array();
+        $result = [];
         $box = range(0, 255);
         $string_length = strlen($string);
 
-        $rnd_key = array();
+        $rnd_key = [];
         for ($i = 0; $i <= 255; $i++) {
             $rnd_key[$i] = $crypt_key[$i % $key_length];
         }
@@ -298,7 +301,7 @@ class Helper
             $box[$j] = $tmp;
         }
 
-        $p1 = $p2 = array();
+        $p1 = $p2 = [];
         for ($a = $j = $i = 0; $i < $string_length; $i++) {
             $a = ($a + 1) % 256;
             $j = ($j + $box[$a]) % 256;
@@ -340,7 +343,7 @@ class Helper
      * @param string $operation encode加密 其他任意字符解密
      * @return string
      */
-    static function encodeParams($str, $key, $operation = 'encode')
+    static function encodeParams(string $str, string $key, string $operation = 'encode'): string
     {
         $result = '';
         static $key_cache;
@@ -390,7 +393,7 @@ class Helper
      * @param string $path_name
      * @return string
      */
-    static function getPath($id, $path_name = '')
+    static function getPath(int $id, string $path_name = ''): string
     {
         $id = (string)abs($id);
         $id = str_pad($id, 9, '0', STR_PAD_LEFT);
@@ -410,9 +413,10 @@ class Helper
      * @param int $timeout
      * @param bool $CA
      * @param string $cacert http://curl.haxx.se/ca/cacert.pem
-     * @return int|mixed|string
+     * @return string
+     * @throws CoreException
      */
-    static function curlRequest($url, $vars = array(), $method = 'POST', $timeout = 10, $CA = false, $cacert = '')
+    static function curlRequest(string $url, $vars = [], string $method = 'POST', int $timeout = 10, bool $CA = false, string $cacert = ''): string
     {
         $method = strtoupper($method);
         $SSL = substr($url, 0, 8) == "https://" ? true : false;
@@ -433,7 +437,7 @@ class Helper
         curl_setopt($ch, CURLOPT_FOLLOWLOCATION, 1);
         curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
         curl_setopt($ch, CURLOPT_CUSTOMREQUEST, $method);
-        curl_setopt($ch, CURLOPT_HTTPHEADER, array("X-HTTP-Method-Override: {$method}"));
+        curl_setopt($ch, CURLOPT_HTTPHEADER, ["X-HTTP-Method-Override: {$method}"]);
 
         if ($SSL && $CA && $cacert) {
             curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, true);
@@ -447,14 +451,12 @@ class Helper
         if ($method == 'POST' || $method == 'PUT') {
             curl_setopt($ch, CURLOPT_POST, 1);
             curl_setopt($ch, CURLOPT_POSTFIELDS, $vars);
-            curl_setopt($ch, CURLOPT_HTTPHEADER, array('Expect:')); //避免data数据过长
+            curl_setopt($ch, CURLOPT_HTTPHEADER, ['Expect:']); //避免data数据过长
         }
         $result = curl_exec($ch);
-        $error_no = curl_errno($ch);
-        if (!$error_no) {
-            $result = trim($result);
-        } else {
-            $result = $error_no;
+        $errorCode = curl_errno($ch);
+        if (!empty($errorCode)) {
+            throw new CoreException("curl请求失败({$errorCode})");
         }
 
         curl_close($ch);
@@ -468,7 +470,7 @@ class Helper
      * @param int $quote_style
      * @return string
      */
-    static function escape($str, $quote_style = ENT_COMPAT)
+    static function escape(string $str, int $quote_style = ENT_COMPAT): string
     {
         return htmlspecialchars($str, $quote_style);
     }
@@ -509,7 +511,7 @@ class Helper
      * @param string $string
      * @return bool
      */
-    static function isChinese($string)
+    static function isChinese(string $string): bool
     {
         if (preg_match("/^[\\x{4e00}-\\x{9fa5}]+$/u", $string)) {
             return true;
@@ -524,7 +526,7 @@ class Helper
      * @param int $mobile
      * @return bool
      */
-    static function isMobile($mobile)
+    static function isMobile(int $mobile): bool
     {
         if (preg_match("/^1[3456789]\\d{9}$/", $mobile)) {
             return true;
@@ -539,9 +541,9 @@ class Helper
      * @param string $id_card
      * @param bool|true $just_check_length 是否只校验长度
      * @return bool
-     * @throws \Exception
+     * @throws Exception
      */
-    static function checkIDCard($id_card, $just_check_length = true)
+    static function checkIDCard(string $id_card, bool $just_check_length = true): bool
     {
         //长度校验
         $length_validate = preg_match('/^([\d]{17}[xX\d]|[\d]{15})$/', $id_card) === 1;
@@ -553,7 +555,7 @@ class Helper
             return false;
         }
 
-        $city_code = array(
+        $city_code = [
             11 => true, 12 => true, 13 => true, 14 => true, 15 => true,
             21 => true, 22 => true, 23 => true,
             31 => true, 32 => true, 33 => true, 34 => true, 35 => true, 36 => true, 37 => true,
@@ -563,7 +565,7 @@ class Helper
             71 => true,
             81 => true, 82 => true,
             91 => true,
-        );
+        ];
 
         //地区校验
         if (!isset($city_code[$id_card[0] . $id_card[1]])) {
@@ -576,23 +578,22 @@ class Helper
                 return null;
             }
 
-            $factor = array(7, 9, 10, 5, 8, 4, 2, 1, 6, 3, 7, 9, 10, 5, 8, 4, 2);
+            $factor = [7, 9, 10, 5, 8, 4, 2, 1, 6, 3, 7, 9, 10, 5, 8, 4, 2];
             //校验码对应值
-            $verify_number_list = array('1', '0', 'X', '9', '8', '7', '6', '5', '4', '3', '2');
+            $verify_number_list = ['1', '0', 'X', '9', '8', '7', '6', '5', '4', '3', '2'];
             $checksum = 0;
             for ($i = 0; $i < 17; $i++) {
                 $checksum += $id_card[$i] * $factor[$i];
             }
 
             $mod = $checksum % 11;
-            $verify_number = $verify_number_list[$mod];
-            return $verify_number;
+            return $verify_number_list[$mod];
         };
 
         $id_card_length = strlen($id_card);
         if ($id_card_length == 15) {
             //超出百岁特殊编码
-            if (array_search(substr($id_card, 12, 3), array('996', '997', '998', '999')) !== false) {
+            if (array_search(substr($id_card, 12, 3), ['996', '997', '998', '999']) !== false) {
                 $id_card = substr($id_card, 0, 6) . '18' . substr($id_card, 6, 9);
             } else {
                 $id_card = substr($id_card, 0, 6) . '19' . substr($id_card, 6, 9);
@@ -608,7 +609,7 @@ class Helper
 
         //校验出生日期
         $birth_day = substr($id_card, 6, 8);
-        $d = new \DateTime($birth_day);
+        $d = new DateTime($birth_day);
         if ($d->format('Y') > date('Y') || $d->format('m') > 12 || $d->format('d') > 31) {
             return false;
         }
@@ -623,17 +624,17 @@ class Helper
      * @param string $op
      * @param string $key
      * @param string $method
-     * @return mixed|string
+     * @return bool|string
      */
-    static function encrypt($data, $op = 'DECODE', $key = '!@#%c*r&o*s^s%p$h~p&', $method = 'AES-256-CBC')
+    static function encrypt($data, string $op = 'DECODE', string $key = '!@#%c*r&o*s^s%p$h~p&', string $method = 'AES-256-CBC')
     {
         $encrypt_key = md5($key);
         if ($op == 'ENCODE') {
             $iv = openssl_random_pseudo_bytes(openssl_cipher_iv_length($method));
             $encrypted = openssl_encrypt($data, $method, $encrypt_key, 0, $iv);
-            $result = str_replace(array('=', '/', '+'), array('', '-', '_'), base64_encode($encrypted . '::' . $iv));
+            $result = str_replace(['=', '/', '+'], ['', '-', '_'], base64_encode($encrypted . '::' . $iv));
         } else {
-            $data = base64_decode(str_replace(array('-', '_'), array('/', '+'), $data));
+            $data = base64_decode(str_replace(['-', '_'], ['/', '+'], $data));
             list($encrypted, $iv) = explode('::', $data);
             $result = openssl_decrypt($encrypted, $method, $encrypt_key, 0, $iv);
         }
@@ -642,27 +643,13 @@ class Helper
     }
 
     /**
-     * 取得用户真实ip
-     *
-     * @return string
-     */
-    static function getIp()
-    {
-        return Request::getInstance()->getClientIPAddress();
-    }
-
-    /**
      * 返回IP的整数形式
      *
      * @param string $ip
-     * @return string
+     * @return int
      */
-    static function getLongIp($ip = '')
+    static function getLongIp(string $ip): int
     {
-        if ($ip == '') {
-            $ip = self::getIp();
-        }
-
         return sprintf("%u", ip2long($ip));
     }
 
@@ -675,7 +662,7 @@ class Helper
      * @param string $suffix
      * @return string
      */
-    static function ftime($time, $format = 'Y-m-d H:i:s', $start_time = 0, $suffix = '前')
+    static function ftime(int $time, string $format = 'Y-m-d H:i:s', int $start_time = 0, string $suffix = '前'): string
     {
         if ($start_time == 0) {
             $start_time = time();
@@ -683,7 +670,7 @@ class Helper
 
         $t = $start_time - $time;
         if ($t < 63072000) {
-            $f = array(
+            $f = [
                 '31536000' => '年',
                 '2592000' => '个月',
                 '604800' => '星期',
@@ -691,7 +678,7 @@ class Helper
                 '3600' => '小时',
                 '60' => '分钟',
                 '1' => '秒'
-            );
+            ];
 
             foreach ($f as $k => $v) {
                 if (0 != $c = floor($t / (int)$k)) {
@@ -709,9 +696,9 @@ class Helper
      * @param int $size
      * @return string
      */
-    static function convert($size)
+    static function convert(int $size): string
     {
-        $unit = array('b', 'kb', 'mb', 'gb', 'tb', 'pb');
+        $unit = ['b', 'kb', 'mb', 'gb', 'tb', 'pb'];
         $s = floor(log($size, 1024));
         $i = (int)$s;
 
