@@ -1,11 +1,14 @@
 <?php
 /**
- * Cross - a micro PHP 5 framework
+ * Cross - a micro PHP framework
  *
  * @link        http://www.crossphp.com
  * @license     MIT License
  */
+
 namespace Cross\DB\SQLAssembler;
+
+use Cross\Exception\CoreException;
 
 /**
  * @author wonli <wonli@live.com>
@@ -15,21 +18,21 @@ namespace Cross\DB\SQLAssembler;
 class MySQLAssembler extends SQLAssembler
 {
     /**
+     * @param string $operator
+     * @param string $field
+     * @param mixed $field_config
+     * @param bool $is_mixed_field
+     * @param string $condition_connector
+     * @param string $connector
+     * @param array $params
+     * @return array
+     * @throws CoreException
      * @see SQLAssembler::parseCondition()
      *
-     * @param $operator
-     * @param $field
-     * @param $field_config
-     * @param $is_mixed_field
-     * @param $condition_connector
-     * @param $connector
-     * @param $params
-     * @return array
-     * @throws \Cross\Exception\CoreException
      */
-    function parseCondition($operator, $field, $field_config, $is_mixed_field, $condition_connector, $connector, array &$params)
+    function parseCondition(string $operator, string $field, $field_config, bool $is_mixed_field, string $condition_connector, string $connector, array &$params): array
     {
-        $condition = array();
+        $condition = [];
         switch ($connector) {
             case 'FIND_IN_SET':
                 $condition[" {$condition_connector} "][] = sprintf('FIND_IN_SET(?, %s)', $field);
