@@ -1,10 +1,11 @@
 <?php
 /**
- * Cross - a micro PHP 5 framework
+ * Cross - a micro PHP framework
  *
  * @link        http://www.crossphp.com
  * @license     MIT License
  */
+
 namespace Cross\I;
 
 /**
@@ -22,7 +23,7 @@ interface SqlInterface
      * @param string|array $where 条件(建议只使用字符串常量,包含变量时请使用数组)
      * @return mixed
      */
-    function get($table, $fields, $where);
+    function get(string $table, string $fields, $where);
 
     /**
      * 批量获取表中的数据
@@ -35,7 +36,7 @@ interface SqlInterface
      * @param int|string $limit 0表示无限制
      * @return mixed
      */
-    function getAll($table, $fields, $where = '', $order = 1, $group_by = 1, $limit = 0);
+    function getAll(string $table, string $fields, $where = [], $order = null, $group_by = null, int $limit = 0);
 
     /**
      * 带分页的查询
@@ -43,12 +44,12 @@ interface SqlInterface
      * @param string $table 表名
      * @param string $fields 字段名
      * @param string|array $where 条件(建议只使用字符串常量,包含变量时请使用数组)
+     * @param array $page ['p', 'limit'] p,当前页 limit,分页条数
      * @param string|int $order 排序
-     * @param array $page array('p', 'limit'); p表示当前页数, limit表示要取出的条数
      * @param int $group_by
      * @return mixed
      */
-    function find($table, $fields, $where, $order = 1, array &$page = array('p', 'limit'), $group_by = 1);
+    function find(string $table, string $fields, $where, array &$page = ['p' => 1, 'limit' => 10], $order = null, $group_by = null);
 
     /**
      * 添加数据
@@ -58,7 +59,7 @@ interface SqlInterface
      * @param bool $multi 是否批量插入
      * @return mixed
      */
-    function add($table, $data, $multi = false);
+    function add(string $table, $data, bool $multi = false);
 
     /**
      * 更新数据
@@ -68,7 +69,7 @@ interface SqlInterface
      * @param string $where 筛选条件
      * @return mixed
      */
-    function update($table, $data, $where);
+    function update(string $table, $data, $where);
 
     /**
      * 删除数据
@@ -77,6 +78,6 @@ interface SqlInterface
      * @param string $where 条件
      * @return mixed
      */
-    function del($table, $where);
+    function del(string $table, $where);
 
 }
