@@ -1,14 +1,15 @@
 <?php
 /**
- * Cross - a micro PHP 5 framework
+ * Cross - a micro PHP framework
  *
  * @link        http://www.crossphp.com
  * @license     MIT License
  */
+
 namespace Cross\Auth;
 
-use Cross\Core\Helper;
 use Cross\I\HttpAuthInterface;
+use Cross\Core\Helper;
 
 /**
  * @author wonli <wonli@live.com>
@@ -24,7 +25,7 @@ class CookieAuth implements HttpAuthInterface
      */
     private $key = '!wl<@>c(r#%o*s&s';
 
-    function __construct($key = '')
+    function __construct(string $key = '')
     {
         if ($key) {
             $this->key = $key;
@@ -37,9 +38,9 @@ class CookieAuth implements HttpAuthInterface
      * @param string $name
      * @param string|array $params
      * @param int $expire
-     * @return bool|mixed
+     * @return bool
      */
-    function set($name, $params, $expire = 0)
+    function set(string $name, $params, int $expire = 0): bool
     {
         if ($params === '' || $params === null) {
             $expire = time() - 3600;
@@ -47,7 +48,7 @@ class CookieAuth implements HttpAuthInterface
         } else {
             $encryptKey = $this->getEncryptKey($name);
             if (is_array($params)) {
-                $params = json_encode($params);
+                $params = json_encode($params, JSON_UNESCAPED_UNICODE);
             }
             $value = Helper::authCode($params, 'ENCODE', $encryptKey);
             if ($expire > 0) {
@@ -60,11 +61,7 @@ class CookieAuth implements HttpAuthInterface
             $cookie_domain = COOKIE_DOMAIN;
         }
 
-        if (setcookie($name, $value, $expire, '/', $cookie_domain, null, true)) {
-            return true;
-        }
-
-        return false;
+        return setcookie($name, $value, $expire, '/', $cookie_domain, null, true);
     }
 
     /**
@@ -74,7 +71,7 @@ class CookieAuth implements HttpAuthInterface
      * @param bool $deCode
      * @return bool|string
      */
-    function get($params, $deCode = false)
+    function get(string $params, bool $deCode = false)
     {
         if (false !== strpos($params, ':') && $deCode) {
             list($name, $arrKey) = explode(':', $params);
@@ -109,14 +106,8 @@ class CookieAuth implements HttpAuthInterface
      * @param string $cookieName
      * @return string
      */
-    protected function getEncryptKey($cookieName)
+    protected function getEncryptKey($cookieName): string
     {
-        if (isset($_SERVER['HTTP_USER_AGENT'])) {
-            $agent = strtolower($_SERVER['HTTP_USER_AGENT']);
-        } else {
-            $agent = 'agent';
-        }
-
-        return md5($agent . $this->key . $cookieName);
+        return md5($cookieName . $this->key . $cookieName);
     }
 }
