@@ -1,26 +1,32 @@
 <?php
 /**
- * Cross - a micro PHP 5 framework
+ * Cross - a micro PHP framework
  *
  * @link        http://www.crossphp.com
  * @license     MIT License
  */
-namespace Cross\DB\Connecter;
+
+namespace Cross\DB\Connector;
 
 use Cross\I\PDOConnecter;
 use PDO;
 
 /**
  * @author wonli <wonli@live.com>
- * Class BaseConnecter
- * @package Cross\DB\Connecter
+ * Class BaseConnector
+ * @package Cross\DB\Connector
  */
-abstract class BaseConnecter implements PDOConnecter
+abstract class BaseConnector implements PDOConnecter
 {
     /**
      * @var PDO
      */
     protected $pdo;
+
+    /**
+     * @var string
+     */
+    protected $sequence;
 
     /**
      * 合并用户输入的options
@@ -29,7 +35,7 @@ abstract class BaseConnecter implements PDOConnecter
      * @param array $options
      * @return array
      */
-    protected static function getOptions(array $default_options, array $options)
+    protected static function getOptions(array $default_options, array $options): array
     {
         if (!empty($options)) {
             foreach ($options as $option_key => $option_val) {
@@ -38,5 +44,15 @@ abstract class BaseConnecter implements PDOConnecter
         }
 
         return $default_options;
+    }
+
+    /**
+     * 设置序号
+     *
+     * @param string $sequence
+     */
+    public function setSequence(string $sequence): void
+    {
+        $this->sequence = $sequence;
     }
 }
