@@ -1,6 +1,6 @@
 <?php
 /**
- * Cross - a micro PHP 5 framework
+ * Cross - a micro PHP framework
  *
  * @link        http://www.crossphp.com
  * @license     MIT License
@@ -8,8 +8,6 @@
 
 namespace Cross\Exception;
 
-use Cross\Http\Response;
-use Exception;
 
 /**
  * @author wonli <wonli@live.com>
@@ -18,12 +16,5 @@ use Exception;
  */
 class CacheException extends CrossException
 {
-    function errorHandler(Exception $e)
-    {
-        $cp_error = $this->cpExceptionSource($e);
-        $code = $e->getCode() ? $e->getCode() : 200;
 
-        Response::getInstance()->setResponseStatus($code)
-            ->display($cp_error, __DIR__ . '/tpl/front_error.tpl.php');
-    }
 }
