@@ -1,6 +1,6 @@
 <?php
 /**
- * Cross - a micro PHP 5 framework
+ * Cross - a micro PHP framework
  *
  * @link        http://www.crossphp.com
  * @license     MIT License
@@ -8,22 +8,15 @@
 
 namespace Cross\Exception;
 
-use Cross\Http\Response;
-use Exception;
 
 /**
+ * 处理逻辑出错时候抛出
+ *
  * @author wonli <wonli@live.com>
  * Class FrontException
  * @package Cross\Exception
  */
 class FrontException extends CrossException
 {
-    function errorHandler(Exception $e)
-    {
-        $cp_error = $this->cpExceptionSource($e);
-        $code = $e->getCode() ? $e->getCode() : 200;
-
-        Response::getInstance()->setResponseStatus($code)
-            ->display($cp_error, __DIR__ . '/tpl/front_error.tpl.php');
-    }
+    protected $httpStatusCode = 400;
 }
