@@ -6,34 +6,36 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>CrossPHP Exception</title>
     <style type="text/css">
-        body {font:14px "12px/1.7 -apple-system","微软雅黑",helvetica,arial,sans-serif;color:#333;background-color:#f5f5f5;padding:0;margin: 0;}
-        a {text-decoration:none;}
-        .header {padding:25px 10px 15px 10px;color:white;background:#a94442;border-top:1px solid #990808;box-shadow:#666 0 1px 1px}
+        body {font:14px "12px/1.7 -apple-system","微软雅黑",helvetica,arial,sans-serif;color:#111;background-color:#fff;padding:0;margin: 0}
+        a {text-decoration:none}
+        .padding {padding-left:15px !important;padding-right:15px !important}
+        .header {padding-top:25px;padding-bottom:15px;color:#262626;font-weight:400;background:#f0f8ff;border-bottom:1px solid #f1f1f1}
         .container{height:100%;width:100%;margin:0;padding:0;left:0;top:0}
-        .title {margin:0;color:#d7b8b8;text-shadow:0 1px 2px rgba(0, 0, 0, .1)}
-        .message {font-size:30px;padding:15px 0;word-wrap:break-word}
-        .stack-container {height:100%;position:relative;border-top:5px solid #be0000}
-        .details {padding:10px}
-        .frame {padding:14px;background:#fff;cursor:pointer}
-        .frame.active {background-color:#666666;color:#f3f3f3}
-        .frame:not(.active):hover {background:#dfdfdf;color:#5d5d5d}
-        .code-container{padding:10px;margin-top:5px;min-height:325px;}
-        .code-segment {padding:5px 0;min-height:325px;background-color:#fff;border:1px solid #f2f2f2;box-shadow:#e5e5ef 0 1px 3px}
-        .code-line span, .code-line code {display:inline-block;}
+        .title {margin:0;color:#818181}
+        .message {font-size:36px;padding:15px 0;word-wrap:break-word}
+        .stack-container {height:100%;position:relative;border-top:6px solid #be0000}
+        .trace-container {padding:0}
+        .frame {padding:15px;background:#fff;cursor:pointer}
+        .frame.active {background-color:#3F51B5;color:#fff}
+        .frame:not(.active):hover {background:rgba(236, 236, 236, 0.66)}
+        .code-container{min-height:325px}
+        .code-segment {min-height:325px;background-color:#fdfdfd;border-bottom:1px solid #f1f1f1}
+        .code-line span, .code-line code {display:inline-block}
         .code-line code {margin-left:-10px}
-        .code-line.active {background-color:#dfdfdf}
-        .data-table-container {padding:10px}
+        .code-line.active {background-color:#ffd2d2}
+        .data-table-container {padding:0}
+        .table-content {width:100%}
         .data-table {width:100%; margin:10px 0}
-        .data-table label {font-size:16px;font-weight:bold;color:#525252;margin:10px 0;padding:10px 0;display:block;border-bottom:1px dotted #b0b0b0;}
+        .data-table label {font-size:16px;font-weight:bold;color:#525252;margin:10px 0;padding:10px 0;display:block;border-bottom:1px dotted #b0b0b0}
         .data-table tbody {font:13px consolas, monospace}
-        .data-table thead {display:none;} .data-table tr {padding:5px 0}
+        .data-table thead {display:none} .data-table tr {padding:5px 0}
         .data-table td:first-child {width:30%;min-width:130px;overflow:hidden;font-weight:bold;color:#463C54;padding-right:5px}
         .data-table td:last-child {width:70%;-ms-word-break:break-all;word-break:break-all;-webkit-hyphens:auto;-moz-hyphens:auto;hyphens:auto}
         .data-table .empty {color:rgba(0, 0, 0, .3);font-style:italic}
         .trace-info{display:none}
         .line{border-right:1px solid #d9d9d9;padding-right:10px;width:50px;text-align:right;line-height:25px}
         .footer{text-align:center;height:30px;background: #333;color:#808080;line-height:30px;box-sizing:border-box;border-top:2px solid rgba(0, 0, 0, 0.35)}
-        .footer>a{color:#565656}@media (max-width: 768px) {h1{font-size:18px} .code-container {display:none} .title{font-size:14px} .message {font-size:20px}}
+        .footer>a{color:#565656}@media (max-width: 768px) {h1{font-size:18px} .code-container {display:none} .title{font-size:14px} .message {font-size:24px}}
     </style>
     <!--[if lt IE 9]><style>.container {min-width:1024px}</style><![endif]-->
 </head>
@@ -41,7 +43,7 @@
 <body>
 <div class="container">
     <div class="stack-container">
-        <div class="header">
+        <div class="header padding">
             <h3 class="title">
                 <?php printf('File: %s Line: %s', $message['main']['show_file'], $message['main']['line']); ?>
             </h3>
@@ -99,12 +101,14 @@
                 </div>
             </div>
         </div>
-        <div class="details">
-            <h1>Exception File</h1>
+        <div class="trace-container">
+            <h1 class="padding">Exception File</h1>
             <div class="frame active" id="frame_active"
-                 onclick="cp_exception.main(null)"><?php echo $message["main"]["show_file"] ?></div>
+                 onclick="cp_exception.main(null)">
+                <?php echo $message["main"]["show_file"] ?>
+            </div>
             <?php if (!empty($message['trace'])) : ?>
-                <h1>Trace</h1>
+                <h1 class="padding">Trace</h1>
                 <?php foreach ($message['trace'] as $kf => $mf) : ?>
                     <div class="frame" onclick="cp_exception.track(<?php echo $kf ?>, this, '')">
                         <?php printf('%s (line: %s)', $mf['show_file'], $mf['line']) ?>
@@ -112,7 +116,7 @@
                 <?php endforeach ?>
             <?php endif ?>
             <?php if (!empty($message['previous_trace'])) : ?>
-                <h1>Previous Trace</h1>
+                <h1 class="padding">Previous Trace</h1>
                 <?php foreach ($message['previous_trace'] as $pkf => $pmf) : ?>
                     <div class="frame" onclick="cp_exception.track(<?php echo $pkf ?>, this, 'previous')">
                         <?php printf('%s (line: %s)', $pmf['show_file'], $pmf['line']) ?>
@@ -122,9 +126,9 @@
         </div>
         <div class="data-table-container" id="data-tables">
             <div class="data-table" id="sg-get-data">
-                <label>GET</label>
+                <label class="padding">GET</label>
                 <?php if (!empty($_GET)) : ?>
-                    <table class="data-table">
+                    <table class="table-content padding">
                         <thead>
                         <tr>
                             <td class="data-table-k">Key</td>
@@ -141,14 +145,14 @@
                         </tbody>
                     </table>
                 <?php else : ?>
-                    <span class="empty">empty</span>
+                    <span class="empty padding">-</span>
                 <?php endif ?>
             </div>
 
             <div class="data-table" id="sg-post-data">
-                <label>POST</label>
+                <label class="padding">POST</label>
                 <?php if (!empty($_POST)) : ?>
-                    <table class="data-table">
+                    <table class="table-content padding">
                         <thead>
                         <tr>
                             <td class="data-table-k">Key</td>
@@ -165,14 +169,14 @@
                         </tbody>
                     </table>
                 <?php else : ?>
-                    <span class="empty">empty</span>
+                    <span class="empty padding">-</span>
                 <?php endif ?>
             </div>
 
             <div class="data-table" id="sg-files">
-                <label>Files</label>
+                <label class="padding">Files</label>
                 <?php if (!empty($_FILES)) : ?>
-                    <table class="data-table">
+                    <table class="table-content padding">
                         <thead>
                         <tr>
                             <td class="data-table-k">Key</td>
@@ -189,15 +193,14 @@
                         </tbody>
                     </table>
                 <?php else : ?>
-                    <span class="empty">empty</span>
+                    <span class="empty padding">-</span>
                 <?php endif ?>
             </div>
 
             <div class="data-table" id="sg-cookies">
-                <label>Cookies</label>
-
+                <label class="padding">Cookies</label>
                 <?php if (!empty($_COOKIE)) : ?>
-                    <table class="data-table">
+                    <table class="table-content padding">
                         <thead>
                         <tr>
                             <td class="data-table-k">Key</td>
@@ -214,14 +217,14 @@
                         </tbody>
                     </table>
                 <?php else : ?>
-                    <span class="empty">empty</span>
+                    <span class="empty padding">-</span>
                 <?php endif ?>
             </div>
 
             <div class="data-table" id="sg-session">
-                <label>Session</label>
+                <label class="padding">Session</label>
                 <?php if (!empty($_SESSION)) : ?>
-                    <table class="data-table">
+                    <table class="table-content padding">
                         <thead>
                         <tr>
                             <td class="data-table-k">Key</td>
@@ -238,7 +241,7 @@
                         </tbody>
                     </table>
                 <?php else : ?>
-                    <span class="empty">empty</span>
+                    <span class="empty padding">-</span>
                 <?php endif ?>
             </div>
         </div>
