@@ -1,11 +1,14 @@
 <?php
 /**
- * Cross - a micro PHP 5 framework
+ * Cross - a micro PHP framework
  *
  * @link        http://www.crossphp.com
  * @license     MIT License
  */
+
 namespace Cross\I;
+
+use PDO;
 
 /**
  * PDO连接类的统一接口
@@ -24,15 +27,15 @@ interface PDOConnecter
      * @param array $options
      * @return mixed
      */
-    static function getInstance($dsn, $user, $password, array $options);
+    static function getInstance(string $dsn, string $user, $password, array $options);
 
     /**
      * 获取表的主键名
      *
      * @param string $table_name
-     * @return mixed
+     * @return string
      */
-    function getPK($table_name);
+    function getPK(string $table_name): string;
 
     /**
      * 获取表的字段信息
@@ -41,14 +44,21 @@ interface PDOConnecter
      * @param bool $fields_map
      * @return mixed
      */
-    function getMetaData($table_name, $fields_map = true);
+    function getMetaData(string $table_name, bool $fields_map = true): array;
+
+    /**
+     * 设置序号（oracle）
+     *
+     * @param string $sequence
+     */
+    function setSequence(string $sequence): void;
 
     /**
      * 返回一个PDO连接对象的实例
      *
      * @return mixed
      */
-    function getPDO();
+    function getPDO(): PDO;
 
     /**
      * 最后插入时自增ID的值
