@@ -1,10 +1,11 @@
 <?php
 /**
- * Cross - a micro PHP 5 framework
+ * Cross - a micro PHP framework
  *
  * @link        http://www.crossphp.com
  * @license     MIT License
  */
+
 namespace Cross\I;
 
 /**
@@ -15,17 +16,17 @@ namespace Cross\I;
 interface RouterInterface
 {
     /**
-     * @return mixed controller
+     * @return string
      */
-    function getController();
+    function getController(): string;
 
     /**
-     * @return mixed action
+     * @return string
      */
-    function getAction();
+    function getAction(): string;
 
     /**
-     * @return mixed params
+     * @return array
      */
-    function getParams();
+    function getParams(): array;
 }
