@@ -5,31 +5,29 @@
 
 namespace app\test\controllers;
 
+use Cross\Exception\CoreException;
 use Cross\MVC\Controller;
 
 class Main extends Controller
 {
     /**
      * 取app配置文件
-     *
-     * @throws \Cross\Exception\CoreException
      */
     function getAppConfig()
     {
-        $this->display($this->config->get('router'));
+        $config = $this->config->get('router');
+        return json_encode($config);
     }
 
     /**
      * 设置app配置文件项
-     *
-     * @throws \Cross\Exception\CoreException
      */
     function setAppConfig()
     {
         $a = $this->params['a'];
         $this->config->set('a', $a);
 
-        $this->display(array('a' => $this->config->get('a')));
+        return json_encode(array('a' => $this->config->get('a')));
     }
 
     /**
@@ -46,7 +44,7 @@ class Main extends Controller
      * 生成连接
      *
      * @return string
-     * @throws \Cross\Exception\CoreException
+     * @throws CoreException
      */
     function makeLink()
     {
@@ -65,7 +63,7 @@ class Main extends Controller
         ));
 
         $this->view->setLinkBase('');
-        return $this->view->link('Main:getUrlSecurityParams', array(
+        return $this->view->url('Main:getUrlSecurityParams', array(
             'p1' => $params['p1'],
             'p2' => $params['p2'],
             'p3' => $params['p3']
@@ -76,7 +74,7 @@ class Main extends Controller
      * 生成加密链接
      *
      * @return string
-     * @throws \Cross\Exception\CoreException
+     * @throws CoreException
      */
     function makeEncryptLink()
     {
@@ -95,18 +93,18 @@ class Main extends Controller
         ));
 
         $this->view->setLinkBase('');
-        return $this->view->slink('Main:getUrlSecurityParams', array(
+        return $this->view->sUrl('Main:getUrlSecurityParams', [
             'p1' => $params['p1'],
             'p2' => $params['p2'],
             'p3' => $params['p3']
-        ));
+        ]);
     }
 
     /**
      * url加密 参数解密
      *
      * @cp_params p1, p2, p3
-     * @throws \Cross\Exception\CoreException
+     * @throws CoreException
      */
     function makeEncryptLinkAndDecryptParams()
     {
@@ -127,7 +125,7 @@ class Main extends Controller
         $action = 'getUrlSecurityParams';
 
         $this->view->setLinkBase('');
-        $url = $this->view->slink(sprintf('%s:%s', $controller, $action), array(
+        $url = $this->view->sUrl(sprintf('%s:%s', $controller, $action), array(
             'p1' => $params['p1'],
             'p2' => $params['p2'],
             'p3' => $params['p3']
@@ -136,17 +134,13 @@ class Main extends Controller
         $custom_router_params[] = $controller;
         $custom_router_params[] = $action;
 
-        switch ($link_type) {
-            case 2:
-                list(, $params_string) = explode('?', $url);
-                break;
-
-            default:
-                $url_array = explode($dot, $url);
-                $params_string = end($url_array);
+        if ($link_type > 2) {
+            list(, $params_string) = explode('?', $url);
+        } else {
+            $url_array = explode($dot, $url);
+            $params_string = end($url_array);
         }
 
-        $result = $this->sParams(true, $params_string);
-        $this->display($result);
+        return $this->sParams(true, $params_string);
     }
 }
